+++ commands/joke.js
@@ -66,28 +66,10 @@ function safe_joke(msg) {
     joke(url, msg);
 }
 
-function eight_ball(msg) {
-    const ans_types = ["Affirmative", "Contrary", "Neutral"]
-    const answers = ["Signs point to yes", "Don't count on it", "Reply hazy, try again later", "You may rely on it", "Outlook not so good", "Very doubtful", "It is certain", "Concentrate and ask again", "Without a doubt", "Don't count on it"];
-    const type = [0, 1, 2, 0, 1, 1, 0, 2, 0, 1]
-    var q = "";
-    var msgtok = msg.content.toLowerCase().split(" ");
-    for (var i = 1; i < msgtok.length; i++) {
-        q += (msgtok[i] + " ");
-    }
-    var ch = randomInteger(0, 9);
-    const embed = new Discord.MessageEmbed()
-        .setColor("#90ee90")
-        .setTitle("DrinkWaterBot Presents Joketopia!")
-        .addFields(
-            { 'name': `#${data['id']}`, value: data['joke'] });
-    msg.channel.send(embed)
-}
 
 module.exports = {
     safe_joke: safe_joke,
     raw_joke: raw_joke,
     chuck_norris: chuck_norris,
-    eight_ball: eight_ball,
     docs: docs
 };
